Use localstorage to persist flow
Using localstore, persist different parts of the state and allow users to refresh the page and safely return to where they were in the flow.

If the user visits the page with a new URL param (e.g. they are trying to deploy a different project), localstorage will clear and the flow will start over.

Closes #28