Fix eventually consistent auth issue
If a user has the Authed-User cookie but no edge state, it's likely that that the corresponding KV key hasn't been set in the local KV namespace for the colo running this worker. To fix this, we'll check for Authed-User in the cookie, _and_ the edge state object being empty, and just refresh the page until it shows up.

While this seems like it could cause a loop, I've been testing this and found that it seems to resolve the problem -- because the token provided by the Authed-User cookie is validated on each request, in theory if anything ends up being invalid here, the Authed-User cookie will be purged if the token/values are incorrect at any point in this chain.

Closes #55