Replace document.cookie check in frontend
The work in 3b2221 removes our ability to read `document.cookie`. This commit changes that functionality in the client and instead checks to the value of an `authed` value in the edge state being passed from the worker.

This functionality is tricky because it only happens in a particular state: when the KV data for an auth key isn't yet available (very rare, since these keys are fresh and often appear instantly), but the worker still knows that the user is authed, via cookie headers.

Instead of passing that value to the client, we now intercept it and set a simple `authed` value which is passed from the `validateCookie` function in the worker down into the edge state read by the React application.